Split schema and add teporary working tests
