+++ fasterid/fasterid.py
@@ -1,14 +1,12 @@
 #!/usr/bin/env python3
-
-from pathlib import Path
 from typing import Annotated
 
 from erdi8 import Erdi8
 from fastapi import Body, Depends, FastAPI, HTTPException
-from pydantic import BaseModel, Field
 from sqlalchemy.orm import Session
 
 from fasterid import models
+from fasterid.schema import ErrorModel, IdModel, RequestModel
 from fasterid.crud import (create_new_mapped_erdi8, create_new_prefix,
                            get_last_erdi8, get_mapped_erdi8, update_last_erdi8)
 from fasterid.database import SessionLocal, engine
@@ -20,7 +18,6 @@ settings = get_settings()
 e8 = Erdi8(settings.erdi8_safe)
 app = FastAPI()
 
-
 # Dependency
 def get_db():
     db = SessionLocal()
@@ -28,35 +25,6 @@ def get_db():
         yield db
     finally:
         db.close()
-
-
-Path(settings.fasterid_filename).touch(exist_ok=True)
-
-
-class RequestModel(BaseModel):
-    prefix: str | None = Field(
-        default="",
-        title="The prefix to be added to the erdi8 string",
-        max_length=settings.fasterid_max_prefix_len,
-    )
-    number: int | None = Field(
-        default=1,
-        title="The number of identifiers that need to be generated",
-        lt=settings.fasterid_max_num + 1,
-    )
-    key: list[str] | None = Field(
-        default=[],
-        title="The keys that need to be mapped to identifiers",
-    )
-
-
-class IdModel(BaseModel):
-    id: list
-
-
-class ErrorModel(BaseModel):
-    detail: str
-
 
 @app.post(
     "/",
